Update place table data

+++ db/migrate/20230608165332_create_places.rb
@@ -4,7 +4,7 @@ class CreatePlaces < ActiveRecord::Migration[7.0]
       t.text :description, null: false
       t.string :photo
       t.string :location, null: false
-      t.decimal :rate, null: false
+      t.integer :rate, null: false
       t.references :owner, null: false, foreign_key: { to_table: :users }, index: true
       t.timestamps
     end
