+++ db/migrate/20230608170046_create_reservations.rb
@@ -5,7 +5,7 @@ class CreateReservations < ActiveRecord::Migration[7.0]
       t.references :place, null: false, foreign_key: { to_table: :places }, index: true
       t.date :start_date, null: false
       t.date :end_date, null: false
-      t.decimal :bill, null: false
+      t.integer :bill, null: false
       t.timestamps
     end
   end
